Fixed invalid handling of send()'s return value.
When sending a request to server (Sphinx::Client_t::sendData()) a return
value from send() function was automatically treated as an error whenever
the value was less or equal to zero (i.e., no data was sent).
Problem is that send() can fail with EAGAIN error which means that the
connection isn't broken, but for some reason (maybe high latency) a sending
of data failed. The correct treatment of this kind of error ought to be a
repeated attempt of sending the data.

git-svn-id: http://svn.code.sf.net/p/cppsphinxclient/code/trunk@23 8419f069-1b38-0410-b4e9-ecff45957251